requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0
altair>=5.2.0
matplotlib>=3.8.0
seaborn>=0.13.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring & Logging
structlog==23.2.0

# Financial Data
yfinance==0.2.28
alpha-vantage==2.3.1

# Social Media APIs
tweepy==4.14.0
newsapi-python==0.2.7 